requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
django-allauth==65.15.0
django-cotton==2.6.2
idna==3.11
pillow==12.1.1
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
uv==0.11.0
